% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_Functions.R
\name{Filt_samples}
\alias{Filt_samples}
\title{Filt_samples function}
\usage{
Filt_samples(
  depth = 5000,
  setType = "complete",
  Taxa_table,
  Metadata,
  cutoff_prop = 0.1,
  parallel = 1
)
}
\arguments{
\item{depth}{The depth that you want to filter samples to. Defaults to 5000}

\item{setType}{The dataset that type you want to return. Currently there are
three options "complete" returns the completed cases dataset (main cohort),
"validation" returns the validation cohort, "whole" returns all samples in the dataset.}

\item{Taxa_table}{The ASV or genus table that is to be filtered.}

\item{Metadata}{The metadata used for filtering.}

\item{cutoff_prop}{The cutoff proportion for fitlering rare taxa.}

\item{parallel}{The number of threads you want to use to filter your data.}
}
\value{
Returns both the filtered metadata and taxa_table in a list. Filtering is done as follows:
Samples are filtered by read depth, rare taxa removed, and then samples filtered based on setType.
}
\description{
This function takes a sequencing depth and whether you want to return the whole dataset, the completed cases dataset, or
the dataset that was used for validation.
}
