% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_Functions.R
\name{get_alpha_divs}
\alias{get_alpha_divs}
\title{get_alpha_divs Function
Takes in metadata and returns the assocaited alpha diversity values for each sample.}
\usage{
get_alpha_divs(Metadata, setType = "complete")
}
\arguments{
\item{Metadata}{Input Metadata that is associated with alpha diversity}

\item{setType}{The set of alpha diversity values you want returned. Can either be "complete"
"validation" or "whole"}
}
\value{
Returns a dataframe containing the sample metadata and its related alpha diversity metrics
}
\description{
get_alpha_divs Function
Takes in metadata and returns the assocaited alpha diversity values for each sample.
}
