% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_Functions.R
\name{get_beta_div}
\alias{get_beta_div}
\title{get_beta_div Function
This function generates a list containing the rarified distance metric and its associated metadata table.}
\usage{
get_beta_div(metric, setType, metadata)
}
\arguments{
\item{metric}{Can either be "w_unifrac" to get a Weighted UniFrac distance matrix or "bray_curtis" to get
a Bray Curtis dissimilarity matrix.}

\item{setType}{Choose between "complete" to get the original dataset. Whole to get all samples... although some will be removed as
they lack a depth of 5000 reads. Or validation to get the validation dataset.}

\item{The}{metadata table associated with the dataset.}
}
\value{
Returns a list containing the "Metadata" and "Distance" matrix
}
\description{
get_beta_div Function
This function generates a list containing the rarified distance metric and its associated metadata table.
}
