% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_Functions.R
\name{get_core_features}
\alias{get_core_features}
\title{get_core_features function}
\usage{
get_core_features(Feat_tab, cutoff)
}
\arguments{
\item{Feat_tab}{A feature table  with rownames as features and colnames and samples}

\item{cutoff}{A precent cutoff for the % of samples}
}
\value{
Returns the rows corresponding to features that are present in above the cutoff % of total samples
}
\description{
This function takes in a feature table and a percent cutoff and returns the rows that correspond to features
that are found in more than the cutoff proportion of samples.
}
