% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_Functions.R
\name{remove_rare_features}
\alias{remove_rare_features}
\title{remove_rare_features function}
\usage{
remove_rare_features(table, cutoff_pro = 0.1, parallel = 1)
}
\arguments{
\item{table}{The feature table that needs to be filtered}

\item{cutoff_pro}{The cutoff proportion used to identify rare features. Defaults to 0.1}

\item{parrellel}{The number of threads used to perform filtering. Defaults to 1.}
}
\value{
Returns the filtered feature table
}
\description{
This function takes in a feature table and removes features that are found in less than the specififed cutoff
}
