% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_Functions.R
\name{run_corn_cob_analysis}
\alias{run_corn_cob_analysis}
\title{run_corn_cob_analysis Function
Runs CornCob differential abundance analysis on the input feature and phyloseq object
Automatically runs DA of the feature while controlling for DNA extraction round and differiental variablility}
\usage{
run_corn_cob_analysis(feature, phylo)
}
\arguments{
\item{feature}{Feature to be tested}

\item{phylo}{The phyloseq object that contains the metadata and ASV table to be tested.}
}
\value{
Returns a list containing the DA results from Corncob
}
\description{
run_corn_cob_analysis Function
Runs CornCob differential abundance analysis on the input feature and phyloseq object
Automatically runs DA of the feature while controlling for DNA extraction round and differiental variablility
}
